+++ db/schema.rb
@@ -126,7 +126,7 @@ ActiveRecord::Schema[8.2].define(version: 2024_05_17_075643) do
 
   create_table "identity_connected_accounts", force: :cascade do |t|
     t.datetime "created_at", null: false
-    t.integer "identity_id", null: false
+    t.bigint "identity_id", null: false
     t.string "provider_name", null: false
     t.string "provider_uid", null: false
     t.datetime "updated_at", null: false
